requirements.txt
# Vector database
faiss-cpu==1.13.2

# Hugging Face
huggingface_hub==0.36.2
sentence-transformers==5.2.2

# LangChain ecosystem (new 1.x series)
langchain==1.2.10
langchain-core==1.2.11
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langchain-huggingface==1.2.0

# Utilities
python-dotenv>=1.0.1

# Document loaders
pdfplumber>=0.11.0
python-docx>=1.1.0
unstructured>=0.15.7

# Web API
flask>=3.0
flask-cors>=4.0